pull request issue#3
